[refactor] 마이페이지 화면 책임 분리
