+++ src/pages/mypage/ui/SocialConnectionsCard.tsx
@@ -1,0 +1,72 @@
+interface ToggleProps {
+   checked: boolean;
+   onChange: () => void;
+   disabled?: boolean;
+}
+
+function Toggle({ checked, onChange, disabled = false }: ToggleProps) {
+   return (
+      <button
+         type="button"
+         role="switch"
+         aria-checked={checked}
+         aria-disabled={disabled}
+         onClick={onChange}
+         disabled={disabled}
+         className={`relative w-10 h-6 rounded-full transition-colors shrink-0 ${checked ? 'bg-primary' : 'bg-border'} ${disabled ? 'cursor-not-allowed opacity-50' : ''}`}
+      >
+         <span
+            className={`absolute top-0.5 left-0.5 size-5 bg-white rounded-full shadow transition-transform ${checked ? 'translate-x-4' : 'translate-x-0'}`}
+         />
+      </button>
+   );
+}
+
+interface SocialConnectionsCardProps {
+   socialConnectionApiAvailable: boolean;
+}
+
+export function SocialConnectionsCard({ socialConnectionApiAvailable }: SocialConnectionsCardProps) {
+   return (
+      <div className="bg-background border border-[rgba(0,0,0,0.1)] rounded-[14px] p-6.25 flex flex-col gap-7.5">
+         <p className="text-heading-3-bold text-foreground">간편 로그인 연결</p>
+         <p className="text-body-2-regular text-muted-foreground">
+            현재 문서 기준으로 연결 계정 조회/변경 API가 정의되지 않아 상태 표시는 제공하지 않습니다.
+         </p>
+         <div className="flex flex-col gap-4">
+            <div className="flex items-center justify-between px-1">
+               <div className="flex items-center gap-5">
+                  <div className="flex items-center gap-2.5">
+                     <div className="border border-border-light rounded-full p-0.5">
+                        <img src="/Icon/Logo/Google.svg" alt="Google" className="size-6" />
+                     </div>
+                     <p className="text-body-2-medium text-foreground">Google 계정 연결</p>
+                  </div>
+                  <span className="bg-[#f1f2f4] text-muted-foreground text-caption-1-regular px-1 py-0.5 rounded">
+                     회원가입 계정
+                  </span>
+               </div>
+               <Toggle checked={false} onChange={() => {}} disabled={!socialConnectionApiAvailable} />
+            </div>
+            <div className="border-t border-border pt-4 flex items-center justify-between px-1">
+               <div className="flex items-center gap-2.5">
+                  <div className="size-7 bg-[#ffde00] rounded-full flex items-center justify-center overflow-hidden">
+                     <img src="/Icon/Logo/Kakao.svg" alt="Kakao" className="size-5" />
+                  </div>
+                  <p className="text-body-2-medium text-muted-foreground">카카오 계정 연결</p>
+               </div>
+               <Toggle checked={false} onChange={() => {}} disabled={!socialConnectionApiAvailable} />
+            </div>
+            <div className="border-t border-border pt-4 flex items-center justify-between px-1">
+               <div className="flex items-center gap-2.5">
+                  <div className="size-7 bg-[#00c73c] rounded-full flex items-center justify-center overflow-hidden p-0.5">
+                     <img src="/Icon/Logo/Naver.svg" alt="Naver" className="size-5" />
+                  </div>
+                  <p className="text-body-2-medium text-muted-foreground">네이버 계정 연결</p>
+               </div>
+               <Toggle checked={false} onChange={() => {}} disabled={!socialConnectionApiAvailable} />
+            </div>
+         </div>
+      </div>
+   );
+}
